AssessmentsModelManager is a globally accessible singleton instance and is no longer a QML type

+++ editor/apps/IngeScape-Assessments/controller/task/taskscontroller.cpp
@@ -14,18 +14,18 @@
 
 #include "taskscontroller.h"
 
+#include <controller/assessmentsmodelmanager.h>
+
 /**
  * @brief Constructor
  * @param modelManager
  * @param jsonHelper
  * @param parent
  */
-TasksController::TasksController(AssessmentsModelManager* modelManager, /*JsonHelper* jsonHelper, */QObject *parent)
+TasksController::TasksController(/*JsonHelper* jsonHelper, */QObject *parent)
     : QObject(parent)
     , _currentExperimentation(nullptr)
     , _selectedTask(nullptr)
-    , _modelManager(modelManager)
-    //, _jsonHelper(jsonHelper)
 {
     // Force ownership of our object, it will prevent Qml from stealing it
     QQmlEngine::setObjectOwnership(this, QQmlEngine::CppOwnership);
@@ -55,10 +55,6 @@ TasksController::~TasksController()
     {
         setcurrentExperimentation(nullptr);
     }
-
-    // Reset pointers
-    _modelManager = nullptr;
-    //_jsonHelper = nullptr;
 }
 
 
@@ -154,7 +150,7 @@ void TasksController::createNewTaskWithIngeScapePlatformFilePath(QString taskNam
  */
 void TasksController::deleteTask(TaskM* task)
 {
-    if ((task != nullptr) && (_currentExperimentation != nullptr))
+    if ((task != nullptr) && (_currentExperimentation != nullptr) && (AssessmentsModelManager::Instance() != nullptr))
     {
         if (task == _selectedTask) {
             setselectedTask(nullptr);
@@ -169,7 +165,7 @@ void TasksController::deleteTask(TaskM* task)
         cass_statement_bind_uuid(cassStatement, 1, task->getCassUuid());
 
         // Execute the query or bound statement
-        CassFuture* cassFuture = cass_session_execute(_modelManager->getCassSession(), cassStatement);
+        CassFuture* cassFuture = cass_session_execute(AssessmentsModelManager::Instance()->getCassSession(), cassStatement);
         CassError cassError = cass_future_error_code(cassFuture);
         if (cassError == CASS_OK)
         {
@@ -190,7 +186,7 @@ void TasksController::deleteTask(TaskM* task)
         cass_statement_bind_uuid(cassStatement, 1, task->getCassUuid());
 
         // Execute the query or bound statement
-        cassFuture = cass_session_execute(_modelManager->getCassSession(), cassStatement);
+        cassFuture = cass_session_execute(AssessmentsModelManager::Instance()->getCassSession(), cassStatement);
         cassError = cass_future_error_code(cassFuture);
         if (cassError == CASS_OK)
         {
@@ -452,7 +448,7 @@ void TasksController::saveModificationsOfIndependentVariableEnum(IndependentVari
  */
 void TasksController::deleteIndependentVariable(IndependentVariableM* independentVariable)
 {
-    if ((independentVariable != nullptr) && (_selectedTask != nullptr))
+    if ((independentVariable != nullptr) && (_selectedTask != nullptr) && (AssessmentsModelManager::Instance() != nullptr))
     {
         // Remove independent variable from DB
         const char* query = "DELETE FROM ingescape.independent_var WHERE id_experimentation = ? AND id_task = ? AND id = ?;";
@@ -462,7 +458,7 @@ void TasksController::deleteIndependentVariable(IndependentVariableM* independen
         cass_statement_bind_uuid(cassStatement, 2, independentVariable->getCassUuid());
 
         // Execute the query or bound statement
-        CassFuture* cassFuture = cass_session_execute(_modelManager->getCassSession(), cassStatement);
+        CassFuture* cassFuture = cass_session_execute(AssessmentsModelManager::Instance()->getCassSession(), cassStatement);
         CassError cassError = cass_future_error_code(cassFuture);
         if (cassError == CASS_OK)
         {
@@ -490,12 +486,12 @@ void TasksController::deleteIndependentVariable(IndependentVariableM* independen
  */
 void TasksController::createNewDependentVariable()
 {
-    if (_selectedTask != nullptr)
+    if ((_selectedTask != nullptr) && (AssessmentsModelManager::Instance() != nullptr))
     {
         qDebug() << "Create a new dependent variable";
 
         CassUuid dependentVarUuid;
-        cass_uuid_gen_time(_modelManager->getCassUuidGen(), &dependentVarUuid);
+        cass_uuid_gen_time(AssessmentsModelManager::Instance()->getCassUuidGen(), &dependentVarUuid);
 
         DependentVariableM* dependentVariable = _insertDependentVariableIntoDB(_selectedTask->getExperimentationCassUuid(), _selectedTask->getCassUuid(), "Dep. Var.", "", "", "");
 
@@ -535,10 +531,10 @@ TaskM* TasksController::_createNewTaskWithIngeScapePlatformFileUrl(QString taskN
 {
     TaskM* task = nullptr;
 
-    if (!taskName.isEmpty() && platformFileUrl.isValid() && (_currentExperimentation != nullptr))
+    if (!taskName.isEmpty() && platformFileUrl.isValid() && (_currentExperimentation != nullptr) && (AssessmentsModelManager::Instance() != nullptr))
     {
         CassUuid taskUuid;
-        cass_uuid_gen_time(_modelManager->getCassUuidGen(), &taskUuid);
+        cass_uuid_gen_time(AssessmentsModelManager::Instance()->getCassUuidGen(), &taskUuid);
 
         const char* query = "INSERT INTO ingescape.task (id_experimentation, id, name, platform_file) VALUES (?, ?, ?, ?);";
         CassStatement* cassStatement = cass_statement_new(query, 4);
@@ -548,7 +544,7 @@ TaskM* TasksController::_createNewTaskWithIngeScapePlatformFileUrl(QString taskN
         cass_statement_bind_string(cassStatement, 3, platformFileUrl.toString().toStdString().c_str());
 
         // Execute the query or bound statement
-        CassFuture* cassFuture = cass_session_execute(_modelManager->getCassSession(), cassStatement);
+        CassFuture* cassFuture = cass_session_execute(AssessmentsModelManager::Instance()->getCassSession(), cassStatement);
         CassError cassError = cass_future_error_code(cassFuture);
         if (cassError == CASS_OK)
         {
@@ -594,10 +590,10 @@ IndependentVariableM* TasksController::_insertIndependentVariableIntoDB(CassUuid
 {
     IndependentVariableM* independentVariable = nullptr;
 
-    if (!variableName.isEmpty())
+    if (!variableName.isEmpty() && (AssessmentsModelManager::Instance() != nullptr))
     {
         CassUuid independentVarUuid;
-        cass_uuid_gen_time(_modelManager->getCassUuidGen(), &independentVarUuid);
+        cass_uuid_gen_time(AssessmentsModelManager::Instance()->getCassUuidGen(), &independentVarUuid);
 
         const char* query = "INSERT INTO ingescape.independent_var (id_experimentation, id_task, id, name, description, value_type, enum_values) VALUES (?, ?, ?, ?, ?, ?, ?);";
         CassStatement* cassStatement = cass_statement_new(query, 7);
@@ -610,7 +606,7 @@ IndependentVariableM* TasksController::_insertIndependentVariableIntoDB(CassUuid
         cass_statement_bind_string(cassStatement, 6, enumValues.join(";").toStdString().c_str());
 
         // Execute the query or bound statement
-        CassFuture* cassFuture = cass_session_execute(_modelManager->getCassSession(), cassStatement);
+        CassFuture* cassFuture = cass_session_execute(AssessmentsModelManager::Instance()->getCassSession(), cassStatement);
         CassError cassError = cass_future_error_code(cassFuture);
         if (cassError == CASS_OK)
         {
@@ -636,7 +632,7 @@ DependentVariableM* TasksController::_insertDependentVariableIntoDB(CassUuid exp
     DependentVariableM* dependentVariable = nullptr;
 
     CassUuid dependentVarUuid;
-    cass_uuid_gen_time(_modelManager->getCassUuidGen(), &dependentVarUuid);
+    cass_uuid_gen_time(AssessmentsModelManager::Instance()->getCassUuidGen(), &dependentVarUuid);
 
     const char* query = "INSERT INTO ingescape.dependent_var (id_experimentation, id_task, id, name, description, agent_name, output_name) VALUES (?, ?, ?, ?, ?, ?, ?);";
     CassStatement* cassStatement = cass_statement_new(query, 7);
@@ -649,7 +645,7 @@ DependentVariableM* TasksController::_insertDependentVariableIntoDB(CassUuid exp
     cass_statement_bind_string(cassStatement, 6, outputName.toStdString().c_str());
 
     // Execute the query or bound statement
-    CassFuture* cassFuture = cass_session_execute(_modelManager->getCassSession(), cassStatement);
+    CassFuture* cassFuture = cass_session_execute(AssessmentsModelManager::Instance()->getCassSession(), cassStatement);
     CassError cassError = cass_future_error_code(cassFuture);
     if (cassError == CASS_OK)
     {
